# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List of categories",
                "operationId": "6ca1297d2f1a55e5ed411a60a091e132",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "body": {
                                                        "type": "string",
                                                        "example": "Some category"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create",
                "operationId": "19b08d7576ebc2bd92c60bc022592da1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "body": {
                                                "type": "string",
                                                "example": "Some category"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "body": {
                                                    "type": "string",
                                                    "example": "Some category"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{category}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show category",
                "operationId": "8ae48eff6f9bce66683e80b5f98a1ab6",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID category",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "body": {
                                                    "type": "string",
                                                    "example": "Some category"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete category",
                "operationId": "aaf29e10dc9518978268f6ecc64325c9",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID category",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 204
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Category"
                ],
                "summary": "Edit category",
                "operationId": "76623091f5297b7b7d6a5cf23aacef29",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID category",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "body": {
                                                "type": "string",
                                                "example": "Edit some category"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "body": {
                                                    "type": "string",
                                                    "example": "Some category"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}